feat: add basic propeller pitch validation
